# lib/api-zod/src/generated/types/healthStatus.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * The Science of Professional Textile Care API
 * OpenAPI spec version: 0.1.0
 */

export interface HealthStatus {
  status: string;
}